+++ schema/schema.go
@@ -18,8 +18,8 @@ type SchemaResponse struct {
 
 type Schema struct {
 	Id        string  `json:"id"`
-	CreatedAt int     `json:"createdAt"`
-	UpdatedAt int     `json:"updatedAt"`
+	CreatedAt int64   `json:"createdAt"`
+	UpdatedAt int64   `json:"updatedAt"`
 	Fields    []Field `json:"fields"`
 }
